requirements.txt
fastapi
uvicorn[standard]
starlette
google-adk
litellm
mcp[cli]>=1.0
firebase-admin
pydantic>=2.0
pydantic-settings
structlog
httpx
pyyaml
python-dotenv
PyJWT[crypto]>=2.8
